First Commit of a Barebones libKSC.
It has some basic functionality and a reasonably decent start at a CMakeLists.txt. There's really not much here and no real testing at all, but it does build and I can use it as a dependency of other projects so it's time to start putting this on GitHub and testing it downstream.

Signed-off-by: ZekeDragon <zeke@kirhut.com>